+++ examples/03-dev/demo/project/src/test-5/index/index.js
@@ -1,0 +1,20 @@
+
+require('base')();
+require('common')();
+
+console.log('test-4/index');
+// require('./js/b')();
+
+document.body.addEventListener('click', e => {
+
+    require.ensure([], require => {
+        require('./js/a')();
+    });
+    require.ensure([], require => {
+        require('./js/b')();
+    });
+    require.ensure([], require => {
+        require('./js/c')();
+    });
+
+}, !1);