+++ examples/03-dev/demo/project/src/test-5/index/js/b.js
@@ -1,0 +1,4 @@
+
+module.exports = _ => {
+    console.log('test-5/index/js/b');
+};