+++ examples/03-dev/demo/project/src/test-5/index/js/c.js
@@ -1,0 +1,8 @@
+
+let b = require('./b');
+
+module.exports = _ => {
+    b();
+
+    console.log('test-5/index/js/c');
+};