+++ examples/03-dev/dist.js
@@ -1,4 +1,4 @@
 
 require('../../util/change_cwd_to')(__dirname + '/demo');
 
-require('../../util/exec')('lila dist test/index');
+require('../../util/exec')('lila dist test-5/index');