+++ examples/03-dev/index-cus.js
@@ -3,4 +3,4 @@ require('../../util/change_cwd_to')(__dirname + '/demo');
 
 require('./cus');
 
-require('../../util/exec')('lila dev test/index');
+require('../../util/exec')('lila dev test-5/index');