+++ examples/03-dev/index.js
@@ -1,4 +1,4 @@
 
 require('../../util/change_cwd_to')(__dirname + '/demo');
 
-require('../../util/exec')('lila dev test-2/index');
+require('../../util/exec')('lila dev test-5/index');